Reverted smallbit to standard real(4) type for compatibility in comparisons with builtins

+++ code/3d_write_graphing_data.f90
@@ -100,7 +100,7 @@ subroutine write_graphing_data( &
 	integer, intent(in) :: dummy_f
 	logical, intent(in) :: diagnostics
 	logical, intent(in) :: isotropic
-	real(dp), intent(in)	:: smallbit
+	real, intent(in)	:: smallbit
 
 	integer, intent(out) :: nplots
